+++ src/barometerdrivers/helpers/util.py
